+++ spec/integration/braintree/transaction_gateway_spec.js
@@ -1861,6 +1861,32 @@ describe('TransactionGateway', function () {
       });
     });
 
+    context('network response code/text', function () {
+      it('returns network response code/text', function (done) {
+        let transactionParams = {
+          amount: '5.00',
+          creditCard: {
+            number: '4111111111111111',
+            expirationDate: '05/12'
+          }
+        };
+
+        specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+          assert.isNull(err);
+          assert.isTrue(response.success);
+          assert.equal(response.transaction.type, 'sale');
+          assert.equal(response.transaction.amount, '5.00');
+          assert.equal(response.transaction.processorResponseCode, '1000');
+          assert.equal(response.transaction.processorResponseType, 'approved');
+          assert.exists(response.transaction.authorizationExpiresAt);
+          assert.equal(response.transaction.networkResponseCode, 'XX');
+          assert.equal(response.transaction.networkResponseText, 'sample network response text');
+
+          done();
+        });
+      });
+    });
+
     context('network transaction id', function () {
       it('support visa', function (done) {
         specHelper.defaultGateway.customer.create({}, function () {
@@ -2124,6 +2150,43 @@ describe('TransactionGateway', function () {
           });
         });
       });
+    });
+
+    context('with paypal here', function () {
+      it('sets PayPalHere attributes on the transaction for auth captures', done =>
+        specHelper.defaultGateway.transaction.find('paypal_here_auth_capture_id', function (err, transaction) {
+          assert.equal(transaction.paymentInstrumentType, PaymentInstrumentTypes.PayPalHere);
+
+          assert.isNotNull(transaction.paypalHereDetails);
+          assert.isNotNull(transaction.paypalHereDetails.authorizationId);
+          assert.isNotNull(transaction.paypalHereDetails.captureId);
+          assert.isNotNull(transaction.paypalHereDetails.invoiceId);
+          assert.isNotNull(transaction.paypalHereDetails.last4);
+          assert.isNotNull(transaction.paypalHereDetails.paymentType);
+          assert.isNotNull(transaction.paypalHereDetails.transactionFeeAmount);
+          assert.isNotNull(transaction.paypalHereDetails.transactionFeeCurrencyIsoCode);
+          assert.isNotNull(transaction.paypalHereDetails.transactionInitiationDate);
+          assert.isNotNull(transaction.paypalHereDetails.transactionUpdatedDate);
+
+          done();
+        })
+      );
+      it('sets PayPalHere attributes on the transaciton for sales', done =>
+        specHelper.defaultGateway.transaction.find('paypal_here_sale_id', function (err, transaction) {
+          assert.isNotNull(transaction.paypalhereDetails);
+          assert.isNotNull(transaction.paypalHereDetails.paymentId);
+
+          done();
+        })
+      );
+      it('sets PayPayHere attributes on the transaction for refunds', done =>
+        specHelper.defaultGateway.transaction.find('paypal_here_refund_id', function (err, transaction) {
+          assert.isNotNull(transaction.paypalHereDetails);
+          assert.isNotNull(transaction.paypalHereDetails.refundId);
+
+          done();
+        })
+      );
     });
 
     context('with apple pay', () =>
@@ -4638,6 +4701,11 @@ describe('TransactionGateway', function () {
           assert.isTrue(info.liabilityShiftPossible);
           assert.equal(info.enrolled, 'Y');
           assert.equal(info.status, 'authenticate_successful');
+          assert.equal(info.cavv, 'somebase64value');
+          assert.equal(info.xid, 'xidvalue');
+          assert.equal(info.eciFlag, '07');
+          assert.equal(info.threeDSecureVersion, '1.0.2');
+          assert.equal(info.dsTransactionId, 'dstxnid');
           done();
         })
       );